Added errors objects and validation

+++ onlineshop/source/server/src/products/products.service.ts
@@ -1,4 +1,4 @@
-import { Injectable } from '@nestjs/common';
+import { HttpException, HttpStatus, Injectable } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
 
@@ -18,18 +18,40 @@ export class ProductsService {
   }
 
   async findOne(id: string): Promise<Product | undefined> {
+    const model = await this.productsRepository.findOne(id);
+    if (!model) {
+      throw new HttpException(`Element with ${id} does not exist`, HttpStatus.NOT_FOUND);
+    }
     return this.productsRepository.findOne(id);
   }
 
   async create(dto: ProductDto): Promise<Product> {
-    const product = await this.productsRepository.save(dto);
-    return product;
+    const size = ['S', 'M', 'L'];
+    if (!size.includes(dto.size.toUpperCase())) {
+      throw new HttpException(`Size value is not valid`, HttpStatus.BAD_REQUEST);
+    }
+    return await this.productsRepository.save(dto);
   }
+
   async remove(id: string): Promise<void> {
+    const model = await this.productsRepository.findOne(id);
+    if (!model) {
+      throw new HttpException(`Element with ${id} does not exist`, HttpStatus.NOT_FOUND);
+    }
     await this.productsRepository.delete(id);
   }
 
-  async update(dto: Product): Promise<Product> {
-    return this.productsRepository.save(dto);
+  async update(dto: ProductDto, id: string): Promise<Product> {
+    const model = await this.productsRepository.findOne(id);
+    if (!model) {
+      throw new HttpException(`Element with ${id} does not exist`, HttpStatus.NOT_FOUND);
+    }
+    model.name = dto.name;
+    model.description = dto.description;
+    model.color = dto.color;
+    model.price = dto.price;
+    model.size = dto.size;
+    model.categoryId = dto.categoryId;
+    return this.productsRepository.save(model);
   }
 }
